feat: first class abort errors
