+++ .gitignore
@@ -7,3 +7,4 @@ coverage
 tsconfig.vitest-temp.json
 website/.vitepress/dist
 website/.vitepress/cache
+legacy
