+++ examples/transport-http_abort.output.txt
@@ -1,0 +1,1 @@
+'This operation was aborted'
