+++ src/layers/6_client/client.test.ts
@@ -59,13 +59,26 @@ describe(`transport`, () => {
       const request = fetch.mock.calls[0]?.[0]
       expect(request?.headers.get(`x-foo`)).toEqual(`bar`)
     })
-
-    test(`sends well formed request`, async ({ fetch, graffle }) => {
+    test(`sends spec compliant request`, async ({ fetch, graffle }) => {
       fetch.mockImplementationOnce(() => Promise.resolve(createResponse({ data: { greetings: `Hello World` } })))
       await graffle.rawString({ document: `query { greetings }` })
       const request = fetch.mock.calls[0]?.[0]
       expect(request?.headers.get(`content-type`)).toEqual(CONTENT_TYPE_JSON)
       expect(request?.headers.get(`accept`)).toEqual(CONTENT_TYPE_GQL)
+    })
+    describe(`signal`, () => {
+      test(`AbortController at instance level works`, async () => {
+        const abortController = new AbortController()
+        const graffle = Graffle.create({
+          schema: endpoint,
+          request: { signal: abortController.signal },
+        })
+        const resultPromise = graffle.rawString({ document: `query { id }` })
+        abortController.abort()
+        await expect(resultPromise).rejects.toThrowErrorMatchingInlineSnapshot(
+          `[AbortError: This operation was aborted]`,
+        )
+      })
     })
   })
   describe(`memory`, () => {
